Add tensorflow datasets dependency (fixes #6)
Tensorflow datasets does not seem to specify its dependency on
a specific tensorflow version. This thus leads to a version being
installed with is incompatible with tfds.

This repository is implemented in tf 1.15.3, thus we need to restrict
the tensorflow datasets version to <4.0.0 in order to prevent dependency
on tf >=2.0.